Using json template engine jsonnet
Currently, we maintain a single yaml file for each of our apps
such as debug-portal, controller, which makes it difficult to be sure
of what we are running on our two estates/clusters, prd-sam and
prd-samtemp.
This change introduces a template for each app (e.g.
templates/samcontrol.jsonnet) and the actual yaml/json files
(prd-sam/configs/yaml(json)) are generated on the fly by build.sh
depending on where the user wants to deploy.
For now, the only thing determined on the fly for different estates is
the images to deploy and that gets configured from templates/images.jsonnet.
It uses open source tool called jsonnet which seems to be
most sophisticated tool available for manipulating json/yaml files.