Update test SAM app to describe deployment strategy
The previous deployment should have failed after hitting one pod.
It did not (both pods got the bad image).

@mayank from the SAM team recommended adding a deployment strategy
to the service to indicate that there should be a maximum of one
unavailable pod during an update.

Rolled back to the 'good' version of the app with this change. A
subsequent change will try to roll to the known bad version with
the new deployment strategy in place.