Update the ceph image override in prd-sam and phx-sam to valid tag
etcd wipe destroyed the existing ceph cluster in prd-sam_ceph.

Now we need to provision a new ceph cluster in prd-sam_ceph, but the operator has changed since the cluster was initially provisioned to apply the tag from the ceph cluster manifest to the container config it creates for Ceph daemons.

Update the tag to the latest qualified build (this is the build running in phx-sam_ceph already).
Apply the override to phx-sam_ceph as well. This is effectively a no-op, but it will help guard against potential future problems until the issue noted in the TODO is addressed.