Add ceph CustomerStoragePool for Gater in PHX
Gater plans to onboard to Stateful SAM in PHX this week. This change adds a CustomerStoragePool for Gater in PHX, which will allow their SAM apps to request block storage provided by LegoStore.

Note: I refactored the conditionals in this manifest a bit to make the code a bit more maintainable as we scale out to more clusters.