Add resource labels to storage components to enable querying across components
As part of our release process, it would be useful to be able to query the status of all components owned by storage. Consistent labeling of our resources will make this simpler.

Apply two new labels to resources:

1. A "team=" label, with a value of either "storage-foundation", "legostore", or "sfstore".
2. A "cloud=storage" label.

In the short term, this will allow the weekly release owner to run commands like:

	`kubectl get pods --all-namespaces -l cloud=storage`

to get a complete listing of pods (and their statuses) in the cluster that are owned by the storage team.

Longer term, we can add tooling around these labels to make it simpler for a release owner to make a go/no-go decision on a rollout.

I also did a bit of cleanup:
* Removed fds-sfms-config.jsonnet -- it is no longer in use anywhere (sfms has the correct prometheus config built-in).
* Removed lvprovisioner.jsonnet -- it is no longer in use anywhere: this was the old in-house local volume provisioner, deprecated and replaced by the open-source provisioner (lv-os-provisioner.jsonnet).
* Removed the "slb.sfdc.net/name" label from fds-svc and ceph-metrics-svc. This label is new deprecated/ignored by SLB, and replaced with annotations.
* Changed the VIP type of fds-svc from dsr to tcp. This will allow SLB to provide TLS termination for the VIP (when supported).