team/Retail_DSS moving to Retail_DFS
As part of a shift to new more coarse grained pools for Cross Cloud Platform,
the DSS team has noted that they are now called DFS. We're creating a team for
them, but not shifting their apps, which are happily running in their existing
finer crained pool. They'll deploy new apps in the new pool (which already
references Retail_DFS) and once they're good we'll tear down both the old pool
and the now-defunct team.