switchboard: use env vars instead of args for config overrides
The arg-based config overrides do not appear to be working. I confirmed the same behavior locally. The arg-based approach is a little easier to read, so it may be worthwhile to figure out why it's not working, but for now I'm just moving to env vars to unblock usage of switchboard.