Allow parallel_build.py to drain stdout and stderr
parallel_build.py is used to launch batches of 10 jsonnet processes, setting up pipes to read from the stdout and stderr output from the processes and then dump the results to stdout of parallel_build.py.

It currently waits for the subprocess to exit before trying to read from stdout. In some environments, this has the effect of causing the pipe to saturate, so the jsonnet process hangs trying to write to stdout until something drains the pipe.

Address this by using [Popen.communicate](https://docs.python.org/3/library/subprocess.html#subprocess.Popen.communicate) which reads from the subprocess pipes until hitting EOF before waiting for the subprocess to terminate.