slb-cleanup-unknownpods should be able to clean up unscheduled Terminating pods as well
slb-cleanup-unknownpods was created to ensure that daemonset rolling updates would not get stuck when encountering pods in the Unknown state. Typically pods enter the Unknown state when they are scheduled to a dead node.

Recently, we've seen several incidents where VIPs have failed because the slb-realsvrcfg daemonset rollout was stuck waiting for Terminating pods that were assigned to dead nodes.

This flights a change to the slb-cleanup-unknownpods script allowing it to force-delete pods stuck in the unscheduled Terminating state. Enabled only in prd-sdc for now.