Introduce new slb-opt-out-vips.libsonnet file
This file will be used as a central location for tracking VIPs (or families of VIPs) that are excluded from monitoring.

We have traditionally had problems tracking why VIPs are suppressed from monitoring. The goal of this change is to introduce a more auditable, reviewable, and commentable form of temporarily suppressing monitoring for a VIP. This should hopefully obviate the need for tweaking the (unreviewed, unaudited, uncommentable) Pagerduty filters.

In a future change, I hope to migrate all of the current Pagerduty-filtered VIPs to this file, and then delete the Pagerduty filters.