Introduce slb-ipvs-artifactory-bootstrap daemonset
We need a way to pull images from artifactory from ipvs nodes. When artifactory points to an SLB VIP, traffic from the IPVS node to artifactory is blackholed.

This daemonset will be used to update the /etc/hosts file to point the artifactory cname to the F5 VIP IP, ensuring that traffic doesn't blackhole.

Yes, this is a massive hack.