Refactor global sla-opt-out vips logic so estates without estate-specific logic still get global opt-out vips
We want the ability to define a set of VIPs (using glob syntax) that are opted-out or marked as very important for all estates where VIP watchdog is deployed.

The previous code only applied the global opted-out VIPs for estates which had estate-specific configuration. This refactor ensures that global vip sets are applied globally (even if the estate doesn't have specific config), and allows for global VIVIPs to be defined in the same manner.