slb-envoy-proxy: Deploy official build image
With https://git.soma.salesforce.com/sdn/slb-nginx-proxy/pull/88, we now build slb-envoy-proxy out of the slb-nginx-proxy repo. Cut over to that build.

Add deployerbot deployment container for the envoy image.

Update the command for the slb-envoy-proxy container to flow through hot_restarter -> runner.sh -> envoy to take advantage of the ability to perform hot reloads when certs rotate.

Update the name of the hot reload marker for envoy to reference envoy instead of nginx.