Monitor availability of CDU servers from SYD
Monitor cdu from syd; Amazon is removing support for Cloud HSMs in SYD, so k4a will no longer be functional there. We want to confirm via our watchdog in SYD that servers in CDU are available to serve requests; we no longer really
"care" about the availability of the k4a servers in SYD, and the service will be stood down in SYD until we are using KMS as the HSM-as-a-service.